crewai_mcp_course/requirements.txt
# Core dependencies for CrewAI with FastMCP integration
crewai>=0.30.0
fastmcp>=0.4.0
python-dotenv>=1.0.0

# LLM and AI dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.8

# HTTP and API dependencies
requests>=2.31.0
httpx>=0.25.0
pydantic>=2.0.0

# Utilities
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional dependencies for advanced features
numpy>=1.24.0
pandas>=2.0.0